WIP: TypeScriptify
Use TypeScript instead of Babel to build Wampy. For strong type checking
and to release the need for maintaining the Wampy @typings as well.

Also remove generated files from git repo.

New way to create Wampy object is to use factory function: createWampy().
This is a backward incompatible change.

It is work in progress because tests do not run yet...